Changed return type

+++ code/types.c
@@ -105,7 +105,7 @@ void setFoodCals(Food *food, int cals)
     food->calories = cals;
 }
 
-float calsIngested(Food food, Food ingested)
+int calsIngested(Food food, Food ingested)
 {
-    return (float)food.calories / 100 * ingested.calories;
+    return food.calories / (float)100 * ingested.calories;
 }
